requirements.txt
django-polymorphic==0.7.1
django-reversion==1.8.6
django-mptt==0.7.4
django-taggit==0.12.2
django-sortedm2m==0.9.5
djangorestframework==3.0.4
django-countries==3.1.1
django-embed-video==0.11
django-exclusivebooleanfield==0.3.0